make InternalProcessorContextMockBuilder top level class
